recomendaciones/serializers.py
from rest_framework import serializers
from .models import Usuario, Libro, Recomendacion

class UsuarioSerializer(serializers.ModelSerializer):
    class Meta:
        model = Usuario
        fields = '__all__'

class LibroSerializer(serializers.ModelSerializer):
    class Meta:
        model = Libro
        fields = '__all__'

class RecomendacionSerializer(serializers.ModelSerializer):
    libro = LibroSerializer(read_only=True)

    class Meta:
        model = Recomendacion
        fields = ['libro', 'score']
